% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_fdb_csv}
\alias{read_fdb_csv}
\title{Read CSV file generated by FDB}
\usage{
read_fdb_csv(file)
}
\arguments{
\item{file}{Path to file}
}
\value{
A data frame.
}
\description{
Read a CSV file generated by Ferjedatabanken.
}

