% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfdb-package.R
\docType{package}
\name{readfdb-package}
\alias{readfdb}
\alias{readfdb-package}
\title{Read FDB Files}
\description{
Read files from \href{http://fdb.triona.no/}{Ferjedatabanken}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmalmedal/readfdb}
  \item Report bugs at \url{https://github.com/hmalmedal/readfdb/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

}
\keyword{package}
