% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_fdb_csv}
\alias{read_fdb_csv}
\title{Read CSV file generated by FDB}
\usage{
read_fdb_csv(file, total = FALSE, unparsed = FALSE)
}
\arguments{
\item{file}{Path to file.}

\item{total}{Extract total values.}

\item{unparsed}{Return unparsed data frame.}
}
\value{
A data frame.
}
\description{
Read a CSV file generated by Ferjedatabanken.
}

